refactor(testing): split testing.rs into directory module
HostHarness infrastructure stays in testing/mod.rs.
Tests extracted to domain files:
  harness_tests.rs (21) — DrawCommand routing, snapshot, nav, overlays
  flow_tests.rs    (3)  — split/new_context layout regression guards

All 24 tests pass. Run targeted:
  cargo test "testing::harness_tests"
  cargo test "testing::flow_tests"
